Finish documentation backed by tests.

+++ model/session/session.base.js
@@ -1,6 +1,7 @@
 import db from '../../db/config'
 import { stringDateToInt } from './session.helper'
 
+
 class SessionBase {
 
   get(args) {
@@ -18,7 +19,7 @@ class SessionBase {
       .then((data) => {
         if (data[0] == null) {
           return new Error(JSON.stringify({
-            key: 'id',
+            arg: 'id',
             code: 'non-existent'
           }))
         } else {
@@ -41,7 +42,7 @@ class SessionBase {
         .then((data) => {
           if (data[0] == null) {
             return new Error(JSON.stringify({
-              key: 'user_id',
+              arg: 'user_id',
               code: 'non-existent'
             }))
           } else {
@@ -62,7 +63,7 @@ class SessionBase {
         })
         .catch(() => {
           return new Error(JSON.stringify({
-            key: 'user_id',
+            arg: 'user_id',
             code: 'non-existent'
           }))
         })
@@ -78,7 +79,7 @@ class SessionBase {
           console.info('List session within date range', data)
           if (data[0] == null) {
             return new Error(JSON.stringify({
-              key: 'date',
+              arg: 'date',
               code: 'non-existent'
             }))
           } else {
@@ -109,43 +110,58 @@ class SessionBase {
     if (data.date) {
       _date = stringDateToInt(data.date)
     }
-    return db.result('UPDATE sessions SET status = COALESCE($2, status), user_id = COALESCE($3, user_id), date = COALESCE($4, date), duration_planned = COALESCE($5, duration_planned), duration_success = COALESCE($6, duration_success), location = COALESCE($7, location), note = COALESCE($8, note) WHERE id = $1 RETURNING id, status, user_id, date, duration_planned, duration_success, location, note', [data.id, data.status, data.user_id, _date, data.duration_planned, data.duration_success, data.location, data.note])
-      .then((result) => {
-        if (result.rowCount === 0) {
+    if (data.id) {
+      return db.result('UPDATE sessions SET status = COALESCE($2, status), user_id = COALESCE($3, user_id), date = COALESCE($4, date), duration_planned = COALESCE($5, duration_planned), duration_success = COALESCE($6, duration_success), location = COALESCE($7, location), note = COALESCE($8, note) WHERE id = $1 RETURNING id, status, user_id, date, duration_planned, duration_success, location, note', [data.id, data.status, data.user_id, _date, data.duration_planned, data.duration_success, data.location, data.note])
+        .then((result) => {
+          if (result.rowCount === 0) {
+            return new Error(JSON.stringify({
+              arg: 'id',
+              code: 'non-existent'
+            }))
+          } else {
+            console.info('Updated session with ID:', result.rows)
+            return result.rows[0]
+          }
+        })
+        .catch(() => {
           return new Error(JSON.stringify({
-            key: 'id',
-            code: 'non-existent'
+            code: 'update-error'
           }))
-        } else {
-          console.info('Updated session with ID:', result.rows)
-          return result.rows[0]
-        }
-      })
-      .catch(() => {
-        return new Error(JSON.stringify({
-          code: 'update-error'
-        }))
-      })
+        })
+    } else {
+      return new Error(JSON.stringify({
+        key: 'id',
+        code: 'unspecified'
+      }))
+    }
   }
 
   delete(id) {
-    return db.result('DELETE FROM sessions WHERE id = $1', id)
-      .then((result) => {
-        if (result.rowCount === 0) {
+    if (id) {
+      return db.result('DELETE FROM sessions WHERE id = $1', id)
+        .then((result) => {
+          if (result.rowCount === 0) {
+            return new Error(JSON.stringify({
+              arg: 'id',
+              code: 'non-existent'
+            }))
+          } else {
+            console.info('Deleted session with ID:', id)
+            return result
+          }
+        })
+        .catch(() => {
           return new Error(JSON.stringify({
-            key: 'id',
-            code: 'non-existent'
+            code: 'deletion-error'
           }))
-        } else {
-          console.info('Deleted session with ID:', id)
-          return result
-        }
-      })
-      .catch(() => {
-        return new Error(JSON.stringify({
-          code: 'deletion-error'
-        }))
-      })
+        })
+    } else {
+      return new Error(JSON.stringify({
+        arg: 'id',
+        code: 'unspecified'
+      }))
+    }
+
   }
 }
